service1/requirements.txt
Flask==3.0.2
joblib==1.3.2
langid==1.1.6
nltk==3.8.1
python-dotenv==1.0.1
Requests==2.31.0
spacy==3.7.4
spacy_language_detection==0.2.1
WTForms==3.1.2
scikit-learn==1.4.1post1